chore: refactor chip.stories from makestyles to styled-component
